Split `DSi_I2CHost` into its own header file
- This will break `#include` cycles that exist in commits I plan to cherry-pick
