+++ src/frontend/qt_sdl/PowerManagement/PowerManagementDialog.cpp
@@ -21,6 +21,7 @@
 
 #include "SPI.h"
 #include "DSi_I2C.h"
+#include "DSi_I2CHost.h"
 #include "NDS.h"
 #include "DSi.h"
 #include "Config.h"
