+++ src/frontend/qt_sdl/ROMManager.cpp
@@ -43,6 +43,7 @@
 #include "SPI.h"
 #include "RTC.h"
 #include "DSi_I2C.h"
+#include "DSi_I2CHost.h"
 #include "FreeBIOS.h"
 
 using std::make_unique;
